ensure test tensor on CPU before converting to numpy
Several unit tests use common routine assert_survs() which
performs a model prediction, converts the result to a numpy
array, and then compares to an expected result. If the model
is on a GPU, then the numpy conversion will fail.

Force the output tensor to the CPU before converting to numpy.