allow override of dataset storage directory (#37)
* allow override of dataset storage directory

The `pycox.datasets` module stores datasets under the installation
directory by default. Add an environment variable to allow a user
to specify a different directory for data storage.

This would be useful to allow sharing a data store among a user's
various python venvs, for example. Or to allow a single pycox
installation to be shared among multiple users, each with their
own private data store.

* simplify dataset storage override